initialize checkpoint manager starting height
