+++ relayer/application_relayer.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"errors"
 	"math/big"
-	"strconv"
 	"time"
 
 	"github.com/ava-labs/avalanchego/ids"
@@ -76,6 +75,7 @@ func newApplicationRelayer(
 	db database.RelayerDatabase,
 	ticker *utils.Ticker,
 	sourceBlockchain config.SourceBlockchain,
+	startingHeight uint64,
 	cfg *config.Config,
 ) (*applicationRelayer, error) {
 	quorum, err := cfg.GetWarpQuorum(relayerID.DestinationBlockchainID)
@@ -97,10 +97,11 @@ func newApplicationRelayer(
 	}
 
 	sub := ticker.Subscribe()
-	checkpointManager := newCheckpointManager(logger, db, sub, relayerID)
+
+	checkpointManager := newCheckpointManager(logger, db, sub, relayerID, startingHeight)
 	checkpointManager.run()
 
-	return &applicationRelayer{
+	ar := applicationRelayer{
 		logger:            logger,
 		metrics:           metrics,
 		network:           network,
@@ -111,7 +112,9 @@ func newApplicationRelayer(
 		signingSubnetID:   signingSubnet,
 		warpQuorum:        quorum,
 		checkpointManager: checkpointManager,
-	}, nil
+	}
+
+	return &ar, nil
 }
 
 func (r *applicationRelayer) relayMessage(
@@ -581,75 +584,6 @@ func (r *applicationRelayer) aggregateSignatures(signatureMap map[int]blsSignatu
 		return nil, set.Bits{}, err
 	}
 	return aggSig, vdrBitSet, nil
-}
-
-//
-// Database access
-//
-
-// Determines the height to process from. There are three cases:
-// 1) The database contains the latest processed block data for the chain
-//   - In this case, we return the maximum of the latest processed block and the configured processHistoricalBlocksFromHeight
-//
-// 2) The database has been configured for the chain, but does not contain the latest processed block data
-//   - In this case, we return the configured processHistoricalBlocksFromHeight
-//
-// 3) The database does not contain any information for the chain.
-//   - In this case, we return the configured processHistoricalBlocksFromHeight if it is set, otherwise
-//     we return the chain head.
-func (r *applicationRelayer) calculateStartingBlockHeight(processHistoricalBlocksFromHeight uint64, currentHeight uint64) (uint64, error) {
-	latestProcessedBlock, err := r.getLatestProcessedBlockHeight()
-	if database.IsKeyNotFoundError(err) {
-		// The database does not contain the latest processed block data for the chain,
-		// use the configured process-historical-blocks-from-height instead.
-		// If process-historical-blocks-from-height was not configured, start from the chain head.
-		if processHistoricalBlocksFromHeight == 0 {
-			return currentHeight, nil
-		}
-		return processHistoricalBlocksFromHeight, nil
-	} else if err != nil {
-		// Otherwise, we've encountered an unknown database error
-		r.logger.Error(
-			"failed to get latest block from database",
-			zap.String("relayerID", r.relayerID.ID.String()),
-			zap.Error(err),
-		)
-		return 0, err
-	}
-
-	// If the database does contain the latest processed block data for the key,
-	// use the max of the latest processed block and the configured start block height (if it was provided)
-	if latestProcessedBlock > processHistoricalBlocksFromHeight {
-		r.logger.Info(
-			"Processing historical blocks from the latest processed block in the DB",
-			zap.String("relayerID", r.relayerID.ID.String()),
-			zap.Uint64("latestProcessedBlock", latestProcessedBlock),
-		)
-		return latestProcessedBlock, nil
-	}
-	// Otherwise, return the configured start block height
-	r.logger.Info(
-		"Processing historical blocks from the configured start block height",
-		zap.String("relayerID", r.relayerID.ID.String()),
-		zap.Uint64("processHistoricalBlocksFromHeight", processHistoricalBlocksFromHeight),
-	)
-	return processHistoricalBlocksFromHeight, nil
-}
-
-// Get the latest processed block height from the database.
-// Note that there may be unrelayed messages in the latest processed block
-// because it is updated as soon as a single message from that block is relayed,
-// and there may be multiple message in the same block.
-func (r *applicationRelayer) getLatestProcessedBlockHeight() (uint64, error) {
-	latestProcessedBlockData, err := r.checkpointManager.database.Get(r.relayerID.ID, database.LatestProcessedBlockKey)
-	if err != nil {
-		return 0, err
-	}
-	latestProcessedBlock, err := strconv.ParseUint(string(latestProcessedBlockData), 10, 64)
-	if err != nil {
-		return 0, err
-	}
-	return latestProcessedBlock, nil
 }
 
 //
