+++ relayer/checkpoint_test.go
@@ -63,7 +63,7 @@ func TestCommitHeight(t *testing.T) {
 		id := database.RelayerID{
 			ID: common.BytesToHash(crypto.Keccak256([]byte(test.name))),
 		}
-		cm := newCheckpointManager(logging.NoLog{}, db, nil, id)
+		cm := newCheckpointManager(logging.NoLog{}, db, nil, id, test.currentMaxHeight)
 		heap.Init(test.pendingHeights)
 		cm.pendingCommits = test.pendingHeights
 		cm.committedHeight = test.currentMaxHeight
